Fix crash when initial composition hasn't been performed
