+++ app/src/main/kotlin/com/hippo/ehviewer/ui/scene/FavoritesScene.kt
@@ -73,7 +73,6 @@ import com.hippo.ehviewer.client.data.FavListUrlBuilder
 import com.hippo.ehviewer.client.ehUrl
 import com.hippo.ehviewer.databinding.SceneFavoritesBinding
 import com.hippo.ehviewer.ui.CommonOperations
-import com.hippo.ehviewer.ui.MainActivity
 import com.hippo.ehviewer.ui.legacy.AddDeleteDrawable
 import com.hippo.ehviewer.ui.legacy.BaseDialogBuilder
 import com.hippo.ehviewer.ui.legacy.FabLayout
@@ -429,8 +428,8 @@ class FavoritesScene : SearchBarScene() {
         )
     }
 
-    override fun onDestroyView() {
-        super.onDestroyView()
+    override fun onRelease() {
+        super.onRelease()
         binding.recyclerView.stopScroll()
         mAdapter = null
         _binding = null
@@ -467,9 +466,6 @@ class FavoritesScene : SearchBarScene() {
                 }
             }
         }
-    }.apply {
-        val compositionContext = (requireActivity() as MainActivity).compositionContext!!
-        setParentCompositionContext(compositionContext)
     }
 
     override fun onSearchViewExpanded() {
